+++ test/widgets/charts/dashboard_health_data_test.dart
@@ -4,6 +4,7 @@ import 'package:lotti/classes/health.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/widgets/charts/dashboard_health_config.dart';
 import 'package:lotti/widgets/charts/dashboard_health_data.dart';
+import 'package:lotti/widgets/charts/utils.dart';
 
 void addDiscreteQuantEntity({
   required DateTime dt,
